backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
apitally==0.11.3
backoff==2.2.1
certifi==2024.8.30
click==8.1.7
dnspython==2.6.1
exceptiongroup==1.2.2
fastapi==0.114.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
motor==3.5.1
pydantic==2.9.1
pydantic_core==2.23.3
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
tenacity==9.0.0
typing_extensions==4.12.2
uvicorn==0.30.6
